[Refactor] Modal, TodoCreateModal 폴더 구조 개선 및 불필요한 주석 제거
